iavf: infrastructure to allow more than 16 queues
Instead of the u32 field, use a bitmap to track the queues mapped to
each vector. Also, refactor iavf_configure_queues() to allow negotiating
an arbitrary number of queues with the PF. We send as many
VIRTCHNL_OP_CONFIG_VSI_QUEUES messages as needed depending on the AQ
buffer size.

Signed-off-by: Ahmed Zaki <ahmed.zaki@intel.com>
Signed-off-by: Przemek Kitszel <przemyslaw.kitszel@intel.com>