xarray: make xa_alloc_cyclic() return 0 on all success cases
Change xa_alloc_cyclic() to return 0 even on wrap-around.
Do the same for xa_alloc_cyclic_irq() and xa_alloc_cyclic_bh().

This will prevent any future bug of treating return of 1 as an error:
	int ret = xa_alloc_cyclic(...)
	if (ret) // currently mishandles ret==1
		goto failure;

If there will be someone interested in when wrap-around occurs,
there is still __xa_alloc_cyclic() that behaves as before.
For now there is no such user.

Suggested-by: Matthew Wilcox <willy@infradead.org>
Link: https://lore.kernel.org/netdev/Z9gUd-5t8b5NX2wE@casper.infradead.org
Signed-off-by: Przemek Kitszel <przemyslaw.kitszel@intel.com>
---
CC: Michal Swiatkowski <michal.swiatkowski@linux.intel.com>
CC: Pierre Riteau <pierre@stackhpc.com>
CC: Andrew Morton <akpm@linux-foundation.org>
CC: linux-fsdevel@vger.kernel.org
CC: linux-mm@kvack.org
CC: linux-kernel@vger.kernel.org
Thanks to Andy and Dave for internal review feedback
CC: Andy Shevchenko <andriy.shevchenko@intel.com>
CC: Dave Hansen <dave.hansen@intel.com>